Move helpers to same file
